+++ adapters/lumi/DJT11LM.py
@@ -1,8 +1,9 @@
 from adapters.adapter_with_battery import AdapterWithBattery
 from devices.switch.selector_switch import SelectorSwitch
+from devices.custom_sensor import CustomSensor
 
 
-class SensorVibration(AdapterWithBattery):
+class DJT11LM(AdapterWithBattery):
     def __init__(self, devices):
         super().__init__(devices)
 
@@ -14,6 +15,7 @@ class SensorVibration(AdapterWithBattery):
         self.switch.set_selector_style(SelectorSwitch.SELECTOR_TYPE_MENU)
 
         self.devices.append(self.switch)
+        self.devices.append(CustomSensor(devices, 'stgth', 'strength', ' (Strength)'))
 
     def handleCommand(self, alias, device, device_data, command, level, color):
         self.switch.handle_command(device_data, command, level, color)